Fix linting issues and failed test
